fix(P0.7-P0.8): canonicalize control plane; job-queue/campaign test coverage
P0.7 - execute_control_plane_action_final (055) is now the single canonical
control-plane entry point. Before this, both supabase/functions/control-api
/index.ts and n8n/08_ml_control_plane.json called execute_control_plane_
action_v2 (047), which only ever implemented 3 of ~18 actions - making the
complete, per-action-RBAC-audited implementation dead code. Audited RBAC
(authorize_action_v2's allowlists are a superset matching _final's own
action set, unlike _v2's mismatched allowlist/implementation), payload/
response shape (identical 6 named params, no caller reshaping needed), and
frontend compatibility (apps/ml-console's control() helper already matches)
before switching. _v2 is kept, not retired - _final delegates 2 actions to
it internally. Documented a known limitation surfaced by the audit:
idempotency_key is required but not itself deduplicated at the dispatcher
level.

P0.8 - turned docs/TEST_MATRIX.md into real, executed tests. 8 of 15
scenarios are in P0's scope (messaging/webhooks/job-queue/observability/
control-plane) and now have real pgTAP + shell coverage; 6 of 15 require
the appointments domain or AI-behavior evaluation, neither touched by any
P0 fix, and are honestly marked out of scope rather than silently skipped
- see docs/P0_TEST_MATRIX_COVERAGE.md for the full mapping.

Building this test coverage surfaced a genuine, previously-undiscovered
CRITICAL defect (RISK-017): core.enqueue_integration_job_v1's p_priority
was SMALLINT, and Postgres does not implicitly coerce a bare integer
literal to smallint during function overload resolution. Every one of this
function's 12 real call sites across 044/045/048/051/052/054 passes a bare
literal (e.g. `...,80,5,now(),...`) and was therefore failing at runtime
with "function does not exist" - undiscoverable by reading the SQL text,
only found by actually executing the migration chain against a real
Postgres engine. Fixed by widening the parameter to INTEGER at the single
function signature (043) rather than patching 12 call sites individually;
the underlying integration_jobs.priority column stays SMALLINT (INTEGER->
SMALLINT on INSERT is always a valid assignment cast - the problem was
specific to function-call argument resolution, not storage).

Full local suite: 78/78 pgTAP assertions passing, plus two real two-
connection concurrency proofs (webhook dedup race, concurrent job-claim
race with zero overlap) - see supabase/tests/local_harness/.

+++ supabase/migrations/043_prod_job_queue.sql
@@ -47,10 +47,25 @@ CREATE TABLE IF NOT EXISTS core.integration_job_attempts (
   provider_response JSONB
 );
 
+-- P0.8 CORRECTION (2026-09-19): p_priority was SMALLINT. Postgres does not
+-- implicitly coerce a bare integer literal (e.g. `80`) to smallint during
+-- function overload resolution (only quoted/"unknown"-typed literals, or an
+-- explicit ::smallint cast, resolve) - only an assignment-level cast, which
+-- does not apply to function-call argument matching. Every one of this
+-- function's 12 call sites across 044/045/048/051/052/054 passes a bare
+-- integer literal for this argument (e.g. `...,80,5,now(),...`), so EVERY
+-- call was failing at runtime with "function does not exist" - discovered
+-- only by actually executing this migration chain against a real Postgres
+-- engine (see supabase/tests/p0/009_p0_8_campaigns_and_media.sql), not by
+-- reading the SQL text. Fixed by widening the parameter to INTEGER (the
+-- underlying core.integration_jobs.priority column stays SMALLINT - integer
+-- -> smallint on INSERT is a normal, always-allowed assignment cast; the
+-- problem was specific to function-call overload resolution, not storage).
+-- This is a single fix at the signature, not 12 individual call-site casts.
 CREATE OR REPLACE FUNCTION core.enqueue_integration_job_v1(
  p_business_id UUID,p_job_type TEXT,p_operation TEXT,p_payload JSONB,
  p_dedupe_key TEXT DEFAULT NULL,p_subject_type TEXT DEFAULT NULL,p_subject_id UUID DEFAULT NULL,
- p_priority SMALLINT DEFAULT 100,p_max_attempts INTEGER DEFAULT 8,p_next_attempt_at TIMESTAMPTZ DEFAULT now(),
+ p_priority INTEGER DEFAULT 100,p_max_attempts INTEGER DEFAULT 8,p_next_attempt_at TIMESTAMPTZ DEFAULT now(),
  p_correlation_id TEXT DEFAULT NULL,p_causal_job_id UUID DEFAULT NULL
 ) RETURNS UUID LANGUAGE plpgsql AS $$
 DECLARE v_id UUID;
